+++ src/main/java/com/ayds/Cloudmerce/model/entity/CartEntity.java
@@ -33,13 +33,13 @@ public class CartEntity {
     private DeliveryType deliveryType;
 
     @Column(name = "payment_method_id", nullable = false)
-    private Integer paymentMethodId;
+    private Long paymentMethodId;
 
     @Column(name = "status_id", nullable = false)
     private Integer statusId;
 
     @Column(name = "user_id", nullable = false)
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "created_at", updatable = false)
     private LocalDateTime createdAt;
